# ubl/docs/blueprint/openapi/ubl.openapi.yaml
openapi: 3.0.3
info: { title: UBL 2.0 Gateway API, version: "2.0.0" }
servers: [ { url: http://localhost:8080 } ]
paths:
...
  /link/validate:
    post:
      summary: Validate link against Membrane (no side effects)
      security: [ { bearerAuth: [] }, { cookieAuth: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LinkCommit' } } } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ValidationResult' } } } }
        "409": { description: rejected, content: { application/json: { schema: { $ref: '#/components/schemas/MembraneError' } } } }
  /link/commit:
    post:
      summary: Validate and append link (single source of truth)
      security: [ { bearerAuth: [] }, { cookieAuth: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LinkCommit' } } } }
      responses:
        "200": { description: committed, content: { application/json: { schema: { $ref: '#/components/schemas/LedgerEntry' } } } }
        "409": { description: rejected, content: { application/json: { schema: { $ref: '#/components/schemas/MembraneError' } } } }
  /ledger/{container_id}/tail:
    get:
      summary: SSE tail for container
      parameters: [ { in: path, name: container_id, required: true, schema: { type: string } } ]
      responses:
        "200":
          description: text/event-stream (SSE)
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                event:
                  value: |
                    event: ledger
                    data: {"container_id":"repo://acme/billing","sequence":42,"link_hash":"...","entry_hash":"..."}
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    cookieAuth: { type: apiKey, in: cookie, name: session }
  schemas:
    PactProof:
      type: object
      properties:
        pact_id: { type: string }
        signatures:
          type: array
          items: { type: string }
    LinkCommit:
      type: object
      required: [version,container_id,expected_sequence,previous_hash,atom_hash,intent_class,physics_delta,author_pubkey,signature]
      properties:
        version: { type: integer, example: 1 }
        container_id: { type: string, example: "workspace://acme/dan-42" }
        expected_sequence: { type: integer, example: 7 }
        previous_hash: { type: string, example: "b3e1..." }
        atom_hash: { type: string, example: "a911..." }
        intent_class: { type: string, enum: [Observation,Conservation,Entropy,Evolution] }
        physics_delta: { type: string, description: "i128 serialized as string", example: "0" }
        pact: { $ref: '#/components/schemas/PactProof' }
        author_pubkey: { type: string }
        signature: { type: string }
    ValidationResult:
      type: object
      properties:
        accepted: { type: boolean }
        error: { $ref: '#/components/schemas/MembraneError' }
    LedgerEntry:
      type: object
      properties:
        container_id: { type: string }
        sequence: { type: integer }
        link_hash: { type: string }
        previous_hash: { type: string }
        entry_hash: { type: string }
        ts_unix_ms: { type: integer }
        metadata: { type: object, additionalProperties: true }
    MembraneError:
      type: object
      properties:
        code:
          type: string
          enum: [InvalidVersion,InvalidSignature,InvalidTarget,RealityDrift,SequenceMismatch,PhysicsViolation,PactViolation,UnauthorizedEvolution]
        message: { type: string }
...